login task

+++ app/Http/Controllers/TasksController.php
@@ -16,13 +16,21 @@ class TasksController extends Controller
     // getでmessages/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        //タスク一覧を取得
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // タスク一覧を作成日時の降順で取得
+            //$tasks = Task::all();
+            $tasks = $user->tasks()->orderBy('user_id', 'desc')->paginate(10);
+            $data = [
+                'tasks' => $tasks,
+            ];
+
+        }
         
         //タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        return view('tasks.index', $data);
     }
 
     /**
@@ -55,14 +63,14 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         
-        // タスクを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        // 認証済みユーザ（閲覧者）の投稿として作成タスクを作成リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
         
         // トップページへリダイレクトさせる
-        return redirect('/');
+        return redirect('/dashboard');
     }
 
     /**
@@ -74,8 +82,12 @@ class TasksController extends Controller
     // getでtasks/idにアクセスされた場合の「取得表示処理」
     public function show($id)
     {
-        //idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            //idの値でタスクを検索して取得
+            $task = Task::findOrFail($id);
+        }
         
         //タスク詳細ビューでそれを表示
         return view('tasks.show', [
@@ -92,8 +104,12 @@ class TasksController extends Controller
     // getでtasks/id/editにアクセスされた場合の「更新画面表示処理」
     public function edit($id)
     {
-        //idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            //idの値でタスクを検索して取得
+            $task = Task::findOrFail($id);
+        }
         
         //タスク編集ビューでそれを表示
         return view('tasks.edit', [
@@ -119,13 +135,15 @@ class TasksController extends Controller
         
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
+        
         // タスクを更新
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        $request->user()->tasks()->update([
+        'status' => $request->status,
+        'content' => $request->content,
+        ]);
         
         // トップページへリダイレクトさせる
-        return redirect('/');
+        return redirect('/dashboard');
     }
 
     /**
@@ -139,10 +157,12 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合はタスク投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
         
         // トップページへリダイレクトさせる
-        return redirect('/');
+        return redirect('/dashboard');
     }
 }
